Moved some components into folders
