+++ src/web/spa/src/components/registration/NewAdminPasswordModal.vue
