+++ src/web/spa/src/components/registration/UserInformationForm.vue
@@ -39,10 +39,10 @@
 
 <script setup>
 import { ref } from 'vue'
-import Input from '../ui/Input.vue'
-import Button from '../ui/Button.vue'
+import Input from '@/components/ui/Input.vue'
+import Button from '@/components/ui/Button.vue'
 import { ArrowNarrowRightIcon } from 'vue-tabler-icons'
-import { formData } from '../utility/forms.js'
+import { formData } from '@/components/utility/forms.js'
 
 const emit = defineEmits(['next'])
 const error = ref('')
